refactor(security): move account security logic to SecurityController
Consolidate security-related functionality by moving account security
features from AccountController to SecurityController for better organization
and maintainability. This includes email and password change functionality.
